#15 refactor: local function to shared singletone
